requirements.txt
fastapi==0.105.0
uvicorn==0.25.0
pymongo==4.6.1
python-dotenv
python-multipart
openpyxl
pandas
numpy
jose
passlib